refactor(Player): 플레이어들을 생성하는 기능을 Player로 이동
