+++ src/main/java/blackjack/domain/player/Players.java
@@ -1,39 +1,24 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 
-public final class Player {
+public class Players {
     
     private static final int PLAYER_MIN_SIZE = 1;
     private static final int PLAYER_MAX_SIZE = 6;
-    private static final int MIN_NAME_LENGTH = 2;
-    private static final int MAX_NAME_LENGTH = 8;
+    private final List<Player> players;
     
-    private final String name;
-    
-    public Player(final String name) {
-        validateName(name);
-        this.name = name;
+    private Players(final List<Player> players) {
+        this.players = players;
     }
     
-    private void validateName(final String name) {
-        if (name == null) {
-            throw new IllegalArgumentException("플레이어의 이름은 null이 될 수 없습니다.");
-        }
-        if (name.length() < MIN_NAME_LENGTH || name.length() > MAX_NAME_LENGTH) {
-            throw new IllegalArgumentException("플레이어의 이름은 %d자 이상, %d자 이하여야 합니다.".formatted(
-                    MIN_NAME_LENGTH, MAX_NAME_LENGTH
-            ));
-        }
-    }
-    
-    public static List<Player> createPlayers(final List<String> playerNames) {
+    public static Players createPlayers(final List<String> playerNames) {
         validatePlayers(playerNames);
-        return playerNames.stream()
+        return new Players(playerNames.stream()
                 .map(Player::new)
-                .toList();
+                .toList());
     }
     
     private static void validatePlayers(final List<String> playerNames) {
@@ -52,7 +37,7 @@ public final class Player {
         return new HashSet<>(playerNames).size() != playerNames.size();
     }
     
-    public String getName() {
-        return name;
+    public List<Player> getPlayers() {
+        return new ArrayList<>(players);
     }
 }
