+++ src/test/java/blackjack/domain/BlackjackJudgeTest.java
@@ -1,5 +1,7 @@
 package blackjack.domain;
 
+import blackjack.domain.player.Player;
+
 import java.util.List;
 import java.util.stream.Stream;
